refactor: input to use password input component
