+++ src/components/auth/Oauth/ClientCredentials/ClientCredentialsContent.tsx
@@ -1,6 +1,4 @@
 import { useState } from "react";
-import { PasswordEyeIcon } from "src/assets/PasswordEyeIcon";
-import { PasswordEyeSlashIcon } from "src/assets/PasswordEyeSlashIcon";
 import { AuthErrorAlert } from "src/components/auth/AuthErrorAlert/AuthErrorAlert";
 import { FormComponent } from "src/components/form";
 import { Button } from "src/components/ui-base/Button";
@@ -40,7 +38,6 @@ export function ClientCredentialsForm({
   explicitScopesRequired,
   buttonVariant,
 }: ClientCredentialsFormProps) {
-  const [show, setShow] = useState(false);
   const [formData, setFormData] = useState<ClientCredentialsFormData>({
     clientSecret: "",
     clientId: "",
@@ -49,8 +46,6 @@ export function ClientCredentialsForm({
   const [submitError, setSubmitError] = useState<string | null>(null);
   const { data: providerInfo } = useProviderInfoQuery(provider);
   const metadataFields = providerInfo?.metadata?.input || [];
-
-  const onToggleShowHide = () => setShow((prevShow) => !prevShow);
 
   const handleChange = (
     event: React.FormEvent<HTMLInputElement | HTMLTextAreaElement>,
@@ -102,31 +97,12 @@ export function ClientCredentialsForm({
           placeholder="Client ID"
           onChange={handleChange}
         />
-        <div style={{ display: "flex", gap: ".5rem" }}>
-          <FormComponent.Input
-            id="clientSecret"
-            name="clientSecret"
-            type={show ? "text" : "password"}
-            placeholder="Client Secret"
-            onChange={handleChange}
-          />
-          <Button
-            type="button"
-            style={{
-              height: "2.5rem",
-              width: "2.5rem",
-              display: "flex",
-              alignItems: "center",
-              justifyContent: "center",
-            }}
-            onClick={onToggleShowHide}
-            variant="ghost"
-          >
-            <span style={{ height: "1.5rem", width: "1.5rem" }}>
-              {show ? <PasswordEyeSlashIcon /> : <PasswordEyeIcon />}
-            </span>
-          </Button>
-        </div>
+        <FormComponent.PasswordInput
+          id="clientSecret"
+          name="clientSecret"
+          placeholder="Client Secret"
+          onChange={handleChange}
+        />
         {explicitScopesRequired && (
           <FormComponent.Textarea
             name="scopes"
