+++ src/components/form/index.ts
@@ -1,7 +1,9 @@
 import { Input } from "./Input";
+import { PasswordInput } from "./PasswordInput";
 import { Textarea } from "./Textarea";
 
 export const FormComponent = {
   Input,
   Textarea,
+  PasswordInput,
 };
